Get all the available devices instead the primary one
Get all the devices with the "GetDevices" dbus call.
Do not use "GetPrimaryDevice" as this is really a private api for the
communication between gpm and gnome-shell
Also, we avoid a bug in the "GetPrimaryDevice" interface about no power device
is detected when AC power is connected to the laptop.