Don't keep GSettingsActions around
They were only kept around so that a notify handler could be disconnected.
Since the GSettings object is the canonical source of that data, it makes more
sense to listen to changes there instead of on the action objects.

That was already done for the one remaining key in the schema. This patch
changes that to rebuild the header when any key changes.